Moving reading methods for bytes and ints to a ReadUtils class Actually using Unsigned ints Adding AC3 audio to mp4

+++ src/main/java/nl/siwoc/mediainfo/MediaInfo.java
@@ -28,6 +28,6 @@ public interface MediaInfo {
 	
 	public String getAudioCodec();
 	
-	public short getAudioChannels();
+	public int getAudioChannels();
 	
 }
