+++ src/main/java/nl/siwoc/mediainfo/qtff/StblBox.java
@@ -18,7 +18,7 @@ package nl.siwoc.mediainfo.qtff;
 
 public class StblBox extends Box {
 	
-    public StblBox(Box parent, int size, byte[] data) throws Exception {
+    public StblBox(Box parent, long size, byte[] data) throws Exception {
 		setType("stbl");
 		setSize(size);
 		setParent(parent);
